create custom hooks for provider
