+++ src/tutorial-module/index.ts
@@ -1,4 +1,4 @@
 export { TutorialModule } from "./training-module/TutorialRoot";
 export * from "./interactive-training/InteractiveTrainingProvider";
-export * from "./interactive-training/useInteractiveTrainingContext";
+export * from "./interactive-training/hooks/useInteractiveTrainingContext";
 export * from "./interactive-training/TrainingWrapper";
