+++ src/tutorial-module/interactive-training/InteractiveTrainingProvider.tsx
@@ -1,18 +1,14 @@
-import React, { createContext, useCallback, useEffect, useMemo, useState } from "react";
-import _ from "lodash";
+import React, { createContext, useMemo } from "react";
 import styled from "styled-components";
 
-import { TrainingModule, TrainingModulePage } from "../../domain/entities/TrainingModule";
-import { getCompositionRoot } from "../../webapp/CompositionRoot";
-import { buildTranslate, TranslatableText } from "../../domain/entities/TranslatableText";
+import { TrainingModulePage } from "../../domain/entities/TrainingModule";
 import { ActionButton } from "../../webapp/components/action-button/ActionButton";
-import { D2Api } from "../../types/d2-api";
 import { Maybe } from "../../types/utils";
-import { useBindEvents } from "./useBindEvents";
+import { useBindEvents } from "./hooks/useBindEvents";
 import "./InteractiveTrainingProvider.css";
-import { bind } from "./useInteractiveTrainingContext";
-import { ContainerConfig, defaultContainerConfig } from "../../domain/entities/Config";
+import { bind } from "./hooks/useInteractiveTrainingContext";
 import { TrainingContainer } from "./TrainingContainer";
+import { useModuleState, useTrainingContent, useTrainingData } from "./hooks/useInteractiveTraining";
 
 type TrainingEventKind = "click" | "focus" | "section";
 
@@ -34,35 +30,10 @@ export const InteractiveTrainingProvider: React.FC<TutorialModuleProps> = props 
     const { baseUrl, locale = "en", events, highlightElementsWithBindings, children } = props;
 
     const { pages, containerConfig, d2Api } = useTrainingData({ baseUrl: baseUrl || "" });
+    const { minimizeTraining, showTraining, isMinimized } = useModuleState();
 
-    const [moduleState, setModuleState] = useState<"default" | "minimized">("minimized");
-    const [contents, setContents] = useState<TranslatableText[]>([]);
+    const { textContent, trigger } = useTrainingContent({ pages, locale, d2Api });
 
-    const pageMap = useMemo(() => _.keyBy(pages, p => p.id), [pages]);
-    const translateMethod = useMemo(() => buildTranslate(locale), [locale]);
-
-    const textContent = useMemo(() => {
-        const translatedContent = contents.reduce((acc, content) => `${acc}\n\n${translateMethod(content)}`, "");
-        return baseUrl ? transformD2DocumentUrls(translatedContent, d2Api.apiPath) : translatedContent;
-    }, [contents, translateMethod, baseUrl, d2Api.apiPath]);
-
-    const trigger = useCallback(
-        (props: { targetIds: string[] }) => {
-            const { targetIds } = props;
-            const targetPages = _(targetIds)
-                .map(targetId => pageMap[targetId])
-                .compact()
-                .value();
-            setContents(targetPages);
-        },
-        [pageMap]
-    );
-
-    const minimizeTraining = useCallback(() => {
-        setModuleState("minimized");
-    }, []);
-
-    const isMinimized = moduleState === "minimized";
     const containerClass = `training-scope ${highlightElementsWithBindings ? "highlight-training-elements" : ""}`;
 
     const contextValue = useMemo(() => ({ pages, trigger, events }), [pages, trigger, events]);
@@ -86,55 +57,13 @@ export const InteractiveTrainingProvider: React.FC<TutorialModuleProps> = props 
             </TrainingContainer>
             {pages.length > 0 && (
                 <ActionButtonContainer hidden={!isMinimized}>
-                    <ActionButton onClick={() => setModuleState("default")} />
+                    <ActionButton onClick={showTraining} />
                 </ActionButtonContainer>
             )}
         </InteractiveTrainingContext.Provider>
     );
 };
 
-type UseTrainingData = { baseUrl: string };
-
-function useTrainingData(props: UseTrainingData) {
-    const { baseUrl } = props;
-    const d2Api = useMemo(() => new D2Api({ baseUrl }), [baseUrl]);
-    const compositionRoot = useMemo(() => getCompositionRoot(d2Api), [d2Api]);
-    const [modules, setModules] = useState<TrainingModule[]>([]);
-    const [containerConfig, setContainerConfig] = useState<ContainerConfig>(defaultContainerConfig);
-
-    const pages = useMemo(() => {
-        if (modules.length === 0) return [];
-        return _(modules)
-            .flatMap(module => module.contents.steps)
-            .flatMap(step => step.pages.filter(({ bindings = [] }) => bindings.length > 0))
-            .value();
-    }, [modules]);
-
-    useEffect(() => {
-        compositionRoot.usecases.modules
-            .list()
-            .then(modules => setModules(modules ?? []))
-            .catch(error => {
-                console.error(`Error fetching modules:`, error);
-                setModules([]);
-            });
-
-        compositionRoot.usecases.config
-            .get()
-            .then(config => setContainerConfig(config.containerConfig))
-            .catch(error => {
-                console.error(`Error fetching container config:`, error);
-                setContainerConfig(defaultContainerConfig);
-            });
-    }, [compositionRoot]);
-
-    return { pages, containerConfig, d2Api };
-}
-
 const ActionButtonContainer = styled.div<{ hidden: boolean }>`
     visibility: ${({ hidden }) => (hidden ? "hidden" : "visible")};
 `;
-
-function transformD2DocumentUrls(content: string, apiBaseUrl: string): string {
-    return content.replace(/\.\.\/..\/(documents\/[^)\s"']+)/g, `${apiBaseUrl}/$1`);
-}
