+++ src/tutorial-module/interactive-training/hooks/useCurrentLocation.ts
